Readme updated
Now the wiki is here, so the link for that is here. I also fixed a typo,
now I list the current version number (It's easier now since I use the
GitHub program), the Planned Features section was moved up below Current
Features (which was renamed from Features), and a Recent Updates section
was added for future use.